Refactor polling functionality and add poll disabling
This commit includes a refactor of the polling functionality in 'edit-after-content-form' and 'Form.vue' to include event-driven state changes. It also introduces visual feedback to the user by disabling the polling button when a poll is open. An icon ('icon-poll-disabled.svg') has been added to indicate when the poll is disabled.