Update dependencies and refactor polling functionality
The refactoring includes changes in the options for the data structures in 'edit-after-content-form' and 'Form.vue'. It also includes a handler for 'getApiPaths' in 'index.ts' and an added feature that disables the polling button when a poll is open in 'edit-after-content-form.vue'. Upgrade to '@devprotocol/clubs-core@3.0.0' was also implemented.