Update feed-after-post-content component and improve voting
Significant refactoring of 'feed-after-post-content' and 'Voting' components has been undertaken. The actual user account is now incorporated into the voting mechanism, which is a major enhancement. The 'Poll' type is now also externalized to a 'types.ts' file which simplifies and structures the components better. The '@devprotocol/clubs-plugin-posts' package has been updated to version 0.8.9.