Add immediate result publication setting for polls
This commit introduces an immediate result publication setting for polls. New checkboxes are added to Voting/Form.vue component, allowing users to choose whether they want to immediately publish the poll results. In the case of no immediate publication, a new Unpublished.vue component is shown, informing the user that the poll results are yet to be published. The Poll type in types.ts is updated to reflect these changes.