Update dependencies and code indentation
This commit includes an update to the "@devprotocol/clubs-plugin-posts" dependency in package.json, increases the indentation of several blocks of code in feed-after-post-content component for readability and also updates the 'sass' dependency in yarn.lock file. These changes help keep the project dependencies up to date and improve code readability.